Update Vendor Day Availability

+++ composeApp/src/commonMain/kotlin/domain/Models/CurrentAppointmentBooking.kt
@@ -6,7 +6,7 @@ import domain.Enums.BookingStatus
 import kotlinx.serialization.Serializable
 
 @Serializable @Parcelize
-data class CurrentAppointmentBooking(var isMobileService: Boolean = false, var serviceId: Int = -1,
+data class CurrentAppointmentBooking(var isMobileService: Boolean = false, var serviceId: Long = -1L,
                                      var serviceTypeId: Int? = -1, var appointmentTime: PlatformTime? = null, var day: Int = -1,
                                      var month: Int = -1, var year: Int = -1, var serviceTypeTherapists: ServiceTypeTherapists? = null,
                                      var description : String = "", var monthName: String = "",
