+++ composeApp/src/commonMain/kotlin/domain/Models/ProductCategory.kt
@@ -4,4 +4,4 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ProductCategory(@SerialName("id") val categoryId: Int? = null, @SerialName("vendor_id") val vendorId: Int? = null, @SerialName("categoryTitle") val categoryTitle: String? = null)
+data class ProductCategory(@SerialName("id") val categoryId: Long? = null, @SerialName("vendor_id") val vendorId: Long? = null, @SerialName("categoryTitle") val categoryTitle: String? = null)
