+++ composeApp/src/commonMain/kotlin/domain/Models/Reviewer.kt
@@ -6,5 +6,5 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable @Parcelize
-data class Reviewer(@SerialName("id") val id: Int? = null, @SerialName("firstname") val firstname: String? = null, @SerialName("lastname") val lastname: String? = null,
+data class Reviewer(@SerialName("id") val id: Long? = null, @SerialName("firstname") val firstname: String? = null, @SerialName("lastname") val lastname: String? = null,
                     @SerialName("imageUrl") val profileImageUrl: String? = null, @SerialName("gender") val gender: String? = null): Parcelable
