+++ composeApp/src/commonMain/kotlin/domain/Models/TherapistReviews.kt
@@ -6,6 +6,6 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable @Parcelize
-data class TherapistReviews(@SerialName("id") val id: Int? = null, @SerialName("therapist_id") val therapistId: Int? = null,
+data class TherapistReviews(@SerialName("id") val id: Long? = null, @SerialName("therapist_id") val therapistId: Long? = null,
                             @SerialName("created_at") val reviewDate: String? = null, @SerialName("reviewText") val reviewText: String? = null,
-                            @SerialName("user_id") val userId: Int? = null, @SerialName("reviewer") val reviewer: Reviewer? = null): Parcelable
+                            @SerialName("user_id") val userId: Long? = null, @SerialName("reviewer") val reviewer: Reviewer? = null): Parcelable
