+++ composeApp/src/commonMain/kotlin/domain/Models/VendorRecommendation.kt
@@ -7,8 +7,8 @@ import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable @Parcelize
-data class VendorRecommendation(@SerialName("id") val recommendationId: Int = -1, @SerialName("vendor_id") val vendorId: Int = -1,
+data class VendorRecommendation(@SerialName("id") val recommendationId: Long = -1, @SerialName("vendor_id") val vendorId: Long = -1,
                                 @SerialName("recommendationType") val recommendationType: String = RecommendationType.Services.toPath(),
-                                @SerialName("description") val description: String = "", @SerialName("product_id") val productId: Int = -1,
-                                @SerialName("service_type_id") val serviceTypeId: Int = -1, @SerialName("service_type") val serviceTypeItem: ServiceTypeItem? = null,
+                                @SerialName("description") val description: String = "", @SerialName("product_id") val productId: Long = -1,
+                                @SerialName("service_type_id") val serviceTypeId: Long = -1, @SerialName("service_type") val serviceTypeItem: ServiceTypeItem? = null,
                                 @SerialName("product") val product: Product? = null, @SerialName("imageUrl") val imageUrl: String = "", val isSelected: Boolean = false): Parcelable
