+++ composeApp/src/commonMain/kotlin/domain/Models/VendorTime.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 @Parcelize
-data class VendorTime(@SerialName("id") val id: Int? = null, @SerialName("time_id") val timeId: Int? = null,
+data class VendorTime(@SerialName("id") val id: Long? = null, @SerialName("time_id") val timeId: Int? = null,
                       @SerialName("platform_time") val platformTime: PlatformTime? = null, val isSelected: Boolean = false): Parcelable
 
 data class VendorTimeUIModel(
